Idea for display enclsure
I need to know if the is the route that we are still planning on taking.
Currently, this design has the display sitting on offsets, with room for
cooling underneath. (plus air vents of 1mm in width)

I know we also talked about adding the pi into the screen enclosure, so
that needs to be decided as well.